Added Nova 2 firmware version v2.0.0
A.k.a. Bluetooth Low Energy Beta Release. This branch will be used for test programs that you can opt-in as a developer. This update also introduces the "Minimum SenseCom Version" that one must have in order to get an update. This is important as we may introduce firmware that uses a new protocol that your SenseCom is not yes compatible with.