Reverted main branch back to v1.3.1
Since older SenseCom versions can still download the binaries, despite not having the back end to support BLE. Nova 2 BLE update can still be downloaded via the 'Beta' Branch.

To avoid this issue in the future, the latest firmware will be hosted on a separate branch. It can still be downloaded from the releases over on the right side.